+++ Modules/Ruleset.UserInfo/Ruleset.UserInfo.psm1
@@ -60,6 +60,7 @@ $PublicScripts = @(
 	"Merge-SDDL"
 	"Split-Principal"
 	"Test-Credential"
+	"Test-SDDL"
 	"Test-UPN"
 )